% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectMySQL5.5.R
\name{connectMySQL5.5}
\alias{connectMySQL5.5}
\title{Connect to MySQL v5.5 via Unix Sock}
\usage{
connectMySQL5.5(dbname, username, password)
}
\arguments{
\item{dbname}{database name as character string}
}
\description{
Connect to MySQL v5.5 via Unix Sock
}
