% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcMySQL5.5.R
\name{dcMySQL5.5}
\alias{dcMySQL5.5}
\title{Disconnect MySQL 5.5 Database}
\usage{
dcMySQL5.5(conn, ..., remove = FALSE)
}
\arguments{
\item{conn}{Connection object}

\item{...}{Additional arguments passed to RMySQL's dbDisconnect function.}

\item{remove}{If TRUE, removes the conn object from the Global Environment.}
}
\description{
Disconnect MySQL 5.5 Database
}
